fix(store): max_page_size did nothing on any SQL store
TaskStoreConfig::max_page_size is documented as capping list, and the book
says so under Design Considerations rather than under any one store. The
four SQL stores hold a connection pool and no config at all, and each
carried its own hardcoded n.min(1000) — the same number TaskStoreConfig
uses as its default, so the configurable bound and the hardcoded one agreed
by coincidence.

Measured with the cap set to 10 against 60 stored tasks and a client asking
for 100: InMemoryTaskStore returned 10 and SqliteTaskStore returned all 60.

That is the third instance of one shape in two sessions — a configurable
bound whose default equals the hardcoded fallback, so nothing looks wrong
until somebody tightens it, which is the person who cared enough to look.
The other two were the gRPC transport's stream_connect_timeout and the
push-config cap that the in-memory store's own correct check was masking.

So the fix is the constant, not just the four call sites: DEFAULT_MAX_PAGE_SIZE
is now one value that TaskStoreConfig's default and every SQL store's default
both read, and each SQL store takes with_max_page_size to reach it. They can
no longer drift apart silently, because there is nothing left to drift.

postgres_store/mod.rs crossed the 500-line ratchet on the way, so its pool
construction and its TaskStore impl each moved to a module of their own —
seams at the connection boundary and the trait boundary, leaving mod.rs as
the type, its constructors and its knobs.

One test, proven able to fail by restoring the literal. The asserted numbers
are chosen so they cannot pass by accident: the client asks for more than
the store holds, so a store ignoring the cap returns everything rather than
a number that merely looks bounded.
