+++ crates/a2a-protocol-server/src/store/mod.rs
@@ -31,7 +31,9 @@ pub mod postgres_store;
 pub mod tenant_postgres_store;
 
 pub use retention::{terminal_states, PurgeReport, RetentionPolicy};
-pub use task_store::{ArtifactDelta, InMemoryTaskStore, TaskStore, TaskStoreConfig};
+pub use task_store::{
+    ArtifactDelta, InMemoryTaskStore, TaskStore, TaskStoreConfig, DEFAULT_MAX_PAGE_SIZE,
+};
 pub use tenant::{TenantAwareInMemoryTaskStore, TenantContext, TenantStoreConfig};
 
 /// Normalizes a status timestamp to the `SQLite` `updated_at` column shape,
