+++ crates/a2a-protocol-server/src/store/tenant_sqlite_store.rs
@@ -64,9 +64,24 @@ use super::tenant::TenantContext;
 #[derive(Debug, Clone)]
 pub struct TenantAwareSqliteTaskStore {
     pool: SqlitePool,
+    /// Largest page `list` will return. See
+    /// [`with_max_page_size`](TenantAwareSqliteTaskStore::with_max_page_size).
+    max_page_size: u32,
 }
 
 impl TenantAwareSqliteTaskStore {
+    /// Caps the page size `list` returns, however large a page is asked for.
+    ///
+    /// Defaults to [`DEFAULT_MAX_PAGE_SIZE`], which explains why this store
+    /// needs its own knob rather than reading [`TaskStoreConfig`].
+    ///
+    /// [`TaskStoreConfig`]: crate::store::TaskStoreConfig
+    /// [`DEFAULT_MAX_PAGE_SIZE`]: crate::store::DEFAULT_MAX_PAGE_SIZE
+    #[must_use]
+    pub const fn with_max_page_size(mut self, max: u32) -> Self {
+        self.max_page_size = max;
+        self
+    }
     /// Opens (or creates) a `SQLite` database and initializes the schema.
     ///
     /// # Errors
@@ -124,7 +139,10 @@ impl TenantAwareSqliteTaskStore {
         .execute(&pool)
         .await?;
 
-        Ok(Self { pool })
+        Ok(Self {
+            pool,
+            max_page_size: crate::store::DEFAULT_MAX_PAGE_SIZE,
+        })
     }
 
     /// Deletes terminal tasks that have outlived `policy`.
@@ -294,7 +312,7 @@ impl TaskStore for TenantAwareSqliteTaskStore {
 
             let page_size = match params.page_size {
                 Some(0) | None => 50_u32,
-                Some(n) => n.min(1000),
+                Some(n) => n.min(self.max_page_size),
             };
 
             let limit = super::pagination::fetch_limit(page_size);
